requirements.txt
azure-functions
mysql-connector-python
